requirements.txt
robotframework==5.0.1
robotframework-faker==5.0.0
robotframework-pythonlibcore==3.0.0
robotframework-requests==0.9.3
robotframework-seleniumlibrary==6.0.0
simple-salesforce==1.12.2